src/pyproject.toml
[tool.poetry]
name = "cappa"
version = "0.1.0"
description = ""
authors = ["azat.zakirov <zakirmalay@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.*"
django-mptt = "0.9.0"
django-tinymce = "2.7.0"
django-modeladmin-reorder = "0.3.1"
django-admin-sortable2 = "0.7.3"
Pillow = "7.1.2"
Unidecode = "1.1.1"
psycopg2 = "2.8.5"
gunicorn = "20.0.4"
django-admin-list-filter-dropdown = "1.0.3"
djangorestframework = "3.11.2"
Django = "1.11.29"
requests = "2.24.0"
pyenchant = "3.2.1"
python-memcached = "1.59"
django-filter = "2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
